ディレクトリの構造を整理＋OppField2を削除し、整えました
