+++ src/component/Game/GraveField.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import "../App.css";
+import "../../App.css";
 import GameCard from "./GameCard";
 export default function GraveField(props) {
   return (
