+++ src/component/Game/OppField.jsx
@@ -1,7 +1,7 @@
 import React from "react";
-import "../App.css";
+import "../../App.css";
 import GameCard from "./GameCard";
-export default function OppField1(props) {
+export default function OppField(props) {
   return (
     <React.Fragment>
       <div className="box-row">
