+++ src/component/Select/SelectCardRow.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import "../App.css";
+import "../../App.css";
 import SelectCard from "./SelectCard";
 export default function SelectCardRow(props) {
   return (
