--- src/reducers/member.js
@@ -1,18 +1,0 @@
-import { ONE_OPP } from "../action/index.js";
-
-const members = (state = Number, action) => {
-  switch (action.type) {
-    case ONE_OPP:
-      state = 1;
-      /*return [
-        {
-          Member: action.countMember,
-        },
-      ];*/
-      break;
-    default:
-      return state;
-  }
-};
-
-export default members;
